requirements.txt
cffi==2.0.0
cryptography==46.0.5
pycparser==2.23
python-dotenv==1.2.1
typing_extensions==4.15.0
